feat(router & path alias): add path alias and react router
