+++ web/package.json
@@ -15,6 +15,8 @@
     "antd": "^5.16.2",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-router-dom": "^6.22.3",
+    "vite-tsconfig-paths": "^4.3.2",
     "zustand": "^4.5.2"
   },
   "devDependencies": {
