+++ web/src/features/router/domain/constants/routing-path.constants.ts
@@ -1,0 +1,5 @@
+export const ROUTING_PATH = Object.freeze({
+  home: '/',
+  login: '/login',
+} as const)
+
