+++ web/src/features/router/domain/types/routing-path.type.ts
@@ -1,0 +1,4 @@
+import { ROUTING_PATH } from '../constants/routing-path.constants'
+
+export type RoutingPath = (typeof ROUTING_PATH)[keyof typeof ROUTING_PATH]
+
